rag-from-scratch/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   openai
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
colorama==0.4.6
    # via tqdm
distro==1.9.0
    # via openai
exceptiongroup==1.2.0
    # via anyio
h11==0.14.0
    # via httpcore
httpcore==1.0.4
    # via httpx
httpx==0.27.0
    # via openai
idna==3.6
    # via
    #   anyio
    #   httpx
numpy==1.26.4
    # via
    #   -r requirements.in
    #   pandas
openai==1.13.3
    # via -r requirements.in
pandas==2.2.1
    # via -r requirements.in
pydantic==2.6.3
    # via openai
pydantic-core==2.16.3
    # via pydantic
pymupdf==1.23.26
    # via -r requirements.in
pymupdfb==1.23.22
    # via pymupdf
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.1
    # via pandas
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
tqdm==4.66.2
    # via openai
typing-extensions==4.10.0
    # via
    #   anyio
    #   openai
    #   pydantic
    #   pydantic-core
tzdata==2024.1
    # via pandas
